+++ console2/src/components/organisms/ProjectRawPayloadModeActivity/index.tsx
@@ -18,7 +18,7 @@
  * =====
  */
 import * as React from 'react';
-import { useEffect, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { Form } from 'semantic-ui-react';
 import { ConcordId, ConcordKey, RequestError } from '../../../api/common';
 import { createOrUpdate as apiCreateOrUpdate, RawPayloadMode } from '../../../api/org/project';
@@ -28,6 +28,7 @@ export interface Props {
     orgName: ConcordKey;
     projectId: ConcordId;
     initialValue?: RawPayloadMode;
+    refresh: () => void; // TODO replace with Context?
 }
 
 const getDescription = (m: RawPayloadMode): string => {
@@ -47,19 +48,27 @@ const getDescription = (m: RawPayloadMode): string => {
     }
 };
 
-export default ({ orgName, projectId, initialValue = RawPayloadMode.DISABLED }: Props) => {
+export default ({ orgName, projectId, initialValue = RawPayloadMode.DISABLED, refresh }: Props) => {
     const [value, setValue] = useState(initialValue);
     const [updating, setUpdating] = useState(false);
     const [error, setError] = useState<RequestError>();
 
+    const didMountRef = useRef(false);
     useEffect(() => {
         const update = async () => {
+            if (!didMountRef.current) {
+                didMountRef.current = true;
+                return;
+            }
+
             try {
                 setUpdating(true);
                 await apiCreateOrUpdate(orgName, {
                     id: projectId,
                     rawPayloadMode: value
                 });
+
+                refresh();
             } catch (e) {
                 setError(e);
             } finally {